server/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.7
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
kombu==5.5.0
msgpack==1.1.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.1
pytesseract==0.3.13
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
qrcode==8.2
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
